tests/functional_testing/requirements-test.txt
pytest==7.4.0
selenium==4.11.2
webdriver-manager==3.8.6
Faker==19.3.0
pytest-html==4.0.2
pytest-xdist==3.3.1
pytest-rerunfailures==12.0
pytest-timeout==2.1.0
requests==2.31.0
python-dotenv==1.0.0
reportlab==4.0.4
python-docx==0.8.11
